# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "tool-kit API",
        "description": "Апи регистрации",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8008",
            "description": "Основной API-сервер"
        }
    ],
    "paths": {
...
        "/api/client/get_orders/{orderId}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Получение заявок текущего пользователя",
                "description": "Если есть orderId - вернет конкретную заявку. Если нет - вернет все заявки",
                "operationId": "50cde1abfb64b47ab2a43d9936c71a88",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID заявки (необязательный параметр)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение заявок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "orders": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "order": {
                                                                "type": "string",
                                                                "example": "Animi veniam aliquid vel quia quam at."
                                                            },
                                                            "address": {
                                                                "type": "string",
                                                                "example": "г. Москва, ул. Пушкина 4б"
                                                            },
                                                            "phone": {
                                                                "type": "string",
                                                                "example": "88005553535"
                                                            },
                                                            "user_id": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-02-03T05:34:02.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2025-02-03T05:34:02.000000Z"
                                                            },
                                                            "get_files": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "id": {
                                                                            "type": "integer",
                                                                            "example": 84
                                                                        },
                                                                        "item_name": {
                                                                            "type": "string",
                                                                            "example": "test.jpg"
                                                                        },
                                                                        "path": {
                                                                            "type": "string",
                                                                            "example": "/storage/app/public/test.jpg"
                                                                        },
                                                                        "orders_id": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "created_at": {
                                                                            "type": "string",
                                                                            "format": "date-time",
                                                                            "example": "2025-02-03T05:34:02.000000Z"
                                                                        },
                                                                        "updated_at": {
                                                                            "type": "string",
                                                                            "format": "date-time",
                                                                            "example": "2025-02-03T05:34:02.000000Z"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нехватка прав для действий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Не хватает прав для действий"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неправильный токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Неправильный токен"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/client/delete_order/{orderId}": {
            "delete": {
                "tags": [
                    "Client"
                ],
                "summary": "Удаление заявки пользователем",
                "operationId": "1fa897480c4c53f2d4dcf0896268ab06",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID заявки для удаления",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1..."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление заявки",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "Заявка успешно удалена"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Проблема с нахождением заявки",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "Заявка не найдена"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Нехватка прав для действий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Не хватает прав для действий"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неправильный токен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "example": "Неправильный токен"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}